requirements.txt
# VoxPolish Dependencies
# Core stack for audio, research, and STT

sounddevice>=0.4.6
numpy>=1.24.0
webrtcvad-wheels>=2.0.10.post1
faster-whisper>=0.10.0
google-genai
python-dotenv>=1.0.0
keyboard>=0.13.5
openwakeword>=0.5.1
onnxruntime>=1.15.0
PyQt6>=6.4.0



